refact: Migrate `Auth` class to `Kirby\Auth\Auth`
