+++ config/aliases.php
@@ -75,6 +75,7 @@ return [
 	// Any of these might be removed at any point in the future
 	'kirby\cms\api'            => 'Kirby\Api\Api',
 	'kirby\cms\asset'          => 'Kirby\Filesystem\Asset',
+	'kirby\cms\auth'           => 'Kirby\Auth\Auth',
 	'kirby\cms\blueprint'      => 'Kirby\Blueprint\Blueprint',
 	'kirby\cms\content'        => 'Kirby\Content\Content',
 	'kirby\cms\dir'            => 'Kirby\Filesystem\Dir',
