+++ src/Auth/Method/CodeMethod.php
@@ -2,9 +2,9 @@
 
 namespace Kirby\Auth\Method;
 
+use Kirby\Auth\Auth;
 use Kirby\Auth\Method;
 use Kirby\Auth\Status;
-use Kirby\Cms\Auth;
 use Kirby\Exception\InvalidArgumentException;
 
 /**
