+++ tests/Auth/Method/PasswordMethodTest.php
@@ -3,9 +3,9 @@
 namespace Kirby\Auth\Method;
 
 use InvalidArgumentException;
+use Kirby\Auth\Auth;
 use Kirby\Auth\Method;
 use Kirby\Auth\Status;
-use Kirby\Cms\Auth;
 use Kirby\Cms\User;
 use Kirby\TestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
