+++ tests/Auth/Method/PasswordResetMethodTest.php
@@ -2,10 +2,10 @@
 
 namespace Kirby\Auth\Method;
 
+use Kirby\Auth\Auth;
 use Kirby\Auth\Method;
 use Kirby\Auth\Methods;
 use Kirby\Auth\Status;
-use Kirby\Cms\Auth;
 use Kirby\Exception\InvalidArgumentException;
 use Kirby\TestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
