--- tests/Cms/Auth/fixtures/logins.json
@@ -1,18 +1,0 @@
-{
-    "by-ip": {
-        "87084f11690867b977a611dd2c943a918c3197f4c02b25ab59": {
-            "time": 9999999999,
-            "trials": 9
-        },
-        "38f0a08519792a17e18a251008f3a116977907f921b0b287c8": {
-            "time": 9999999999,
-            "trials": 10
-        }
-    },
-    "by-email": {
-        "homer@simpsons.com": {
-            "time": 9999999999,
-            "trials": 10
-        }
-    }
-}
