+++ tests/bootstrap.php
@@ -18,7 +18,6 @@ require_once __DIR__ . '/Api/mocks.php';
 require_once __DIR__ . '/Auth/mocks.php';
 require_once __DIR__ . '/Cache/mocks.php';
 require_once __DIR__ . '/Cms/mocks.php';
-require_once __DIR__ . '/Cms/Auth/mocks.php';
 require_once __DIR__ . '/Cms/System/mocks.php';
 require_once __DIR__ . '/Content/mocks.php';
 require_once __DIR__ . '/Data/mocks.php';
